refactor(api): create and use classic api handlers
