+++ pages/api/users/index.ts
@@ -1,32 +1,11 @@
-import { NextApiHandler, NextApiRequest as Req, NextApiResponse as Res } from 'next';
-import { User, UserWithoutId } from '@packages/data/user';
-import db from '@packages/utils/db';
-
-export type PullAllResponse = User[];
-export type AddResponse = { _id: string };
+import { NextApiHandler } from 'next';
+import { handlerWithoutId } from '@packages/utils/api';
+import { User } from '@packages/data/user';
 
 const collectionName = 'users';
 
-const pullAllHandler = async (req: Req, res: Res) => {
-  const response = await db.find<User>(collectionName);
-
-  res.status(200).json(response);
-};
-
-const addHandler = async (req: Req, res: Res) => {
-  const { name, email, image, roles }: UserWithoutId = req.body;
-  const newId = await db.insert<UserWithoutId>(collectionName, { name, email, image, roles });
-  const response: AddResponse = { _id: newId };
-
-  res.status(201).json(response);
-};
-
 const handler: NextApiHandler = async (req, res) => {
-  if (req.method === 'GET') {
-    await pullAllHandler(req, res);
-  } else if (req.method === 'POST') {
-    await addHandler(req, res);
-  }
+  await handlerWithoutId<User>(collectionName, req, res);
 };
 
 export default handler;
